split challenge types into their own module
Challenge handling is about to grow per-challenge type state, handles,
and authorization response types. Give that work a home by moving the
existing challenge types out of types.rs into a new challenges module,
keeping the existing names re-exported from the crate root.
