+++ src/challenges.rs
@@ -1,0 +1,86 @@
+//! ACME challenges and challenge type specific handling
+
+use std::borrow::Cow;
+
+use serde::Deserialize;
+
+use crate::types::Problem;
+
+/// An ACME challenge as described in RFC 8555 (section 7.1.5)
+///
+/// <https://datatracker.ietf.org/doc/html/rfc8555#section-7.1.5>
+#[derive(Debug, Deserialize)]
+pub struct Challenge {
+    /// Type of challenge
+    pub r#type: ChallengeType,
+    /// Challenge identifier
+    pub url: String,
+    /// Token for this challenge
+    ///
+    /// Unknown `ChallengeType` instances may omit this field, leaving it empty.
+    #[serde(default)]
+    pub token: String,
+    /// Current status
+    pub status: ChallengeStatus,
+    /// Potential error state
+    pub error: Option<Problem>,
+}
+
+/// The challenge type
+#[allow(missing_docs)]
+#[non_exhaustive]
+#[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
+pub enum ChallengeType {
+    #[serde(rename = "http-01")]
+    Http01,
+    #[serde(rename = "dns-01")]
+    Dns01,
+    #[serde(rename = "tls-alpn-01")]
+    TlsAlpn01,
+    /// Note: Device attestation support is experimental
+    #[serde(rename = "device-attest-01")]
+    DeviceAttest01,
+    #[serde(untagged)]
+    Unknown(String),
+}
+
+/// Status of an ACME [Challenge]
+#[allow(missing_docs)]
+#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub enum ChallengeStatus {
+    Pending,
+    Processing,
+    Valid,
+    Invalid,
+}
+
+/// Attestation payload used for device-attest-01
+///
+/// See <https://datatracker.ietf.org/doc/draft-acme-device-attest/> for details.
+pub struct DeviceAttestation<'a> {
+    /// CBOR encoded attestation payload
+    pub att_obj: Cow<'a, [u8]>,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // https://datatracker.ietf.org/doc/html/rfc8555#section-8.4
+    #[test]
+    fn challenge() {
+        const CHALLENGE: &str = r#"{
+          "type": "dns-01",
+          "url": "https://example.com/acme/chall/Rg5dV14Gh1Q",
+          "status": "pending",
+          "token": "evaGxfADs6pSRb2LAv9IZf17Dt3juxGJ-PCt92wr-oA"
+        }"#;
+
+        let obj = serde_json::from_str::<Challenge>(CHALLENGE).unwrap();
+        assert_eq!(obj.r#type, ChallengeType::Dns01);
+        assert_eq!(obj.url, "https://example.com/acme/chall/Rg5dV14Gh1Q");
+        assert_eq!(obj.status, ChallengeStatus::Pending);
+        assert_eq!(obj.token, "evaGxfADs6pSRb2LAv9IZf17Dt3juxGJ-PCt92wr-oA");
+    }
+}
