+++ src/lib.rs
@@ -45,6 +45,9 @@ use serde::Serialize;
 mod account;
 pub use account::{Account, AccountBuilder, ExternalAccountKey, Key};
 
+pub mod challenges;
+pub use challenges::{Challenge, ChallengeStatus, ChallengeType, DeviceAttestation};
+
 mod crypto;
 pub use crypto::{
     CryptoProvider, HmacKey, HmacKeyProvider, Sha256, SigningKey, SigningKeyProvider,
@@ -59,10 +62,9 @@ pub use order::{
 mod types;
 pub use types::{
     AccountCredentials, Authorization, AuthorizationState, AuthorizationStatus,
-    AuthorizedIdentifier, CertificateIdentifier, Challenge, ChallengeStatus, ChallengeType,
-    DeviceAttestation, EcCurve, Error, Identifier, Jwk, JwkThumbFields, LetsEncrypt, NewAccount,
-    NewOrder, OctetKeyCurve, OrderState, OrderStatus, Problem, ProfileMeta, RevocationReason,
-    RevocationRequest, SigningAlgorithm, Subproblem, ZeroSsl,
+    AuthorizedIdentifier, CertificateIdentifier, EcCurve, Error, Identifier, Jwk, JwkThumbFields,
+    LetsEncrypt, NewAccount, NewOrder, OctetKeyCurve, OrderState, OrderStatus, Problem, ProfileMeta,
+    RevocationReason, RevocationRequest, SigningAlgorithm, Subproblem, ZeroSsl,
 };
 use types::{Directory, JoseJson, Signer};
 #[cfg(feature = "time")]
